requirements.txt
python-docx 
csv2pdf
torch
openai-whisper